⚡ SPRINT 1: Legal Foundation + Numerical Compliance
P0-1 LICENSE MISMATCH (RESOLVED):
- Updated 15 files from MIT to Proprietary
- pyproject.toml, Cargo.toml, README badge, CITATION.cff
- All Rust crates (glass_cockpit, global_eye, hyper_core, hyper_bridge)
- SDK (qtt-sdk), CUDA setup.py, .zenodo.json
- tensornet/site/generator.py default

P0-2 FLOAT64 COMPLIANCE (RESOLVED - Phase 11-15):
- tensornet/medical/hemo.py: velocity, pressure, geometry → float64
- tensornet/racing/wake.py: wake_field, turbulence → float64
- tensornet/defense/ballistics.py: gravity tensor → float64
- tensornet/emergency/fire.py: fuel, heat, elevation → float64
- tensornet/agri/microclimate.py: temperature, humidity, co2, velocity → float64
- tensornet/fusion/tokamak.py: particle initialization → float64
- tensornet/cyber/grid_shock.py: pressure, capacity → float64

P0-4 REPRODUCIBILITY (RESOLVED):
- tensornet/fusion/tokamak.py: Added seed=42 param to create_plasma()
- tensornet/cyber/grid_shock.py: Added torch.manual_seed(42) + np.random.seed(42)
- tensornet/emergency/fire.py: Added step-based seeding for spotting

P2-3 RUFF VERSION (RESOLVED):
- .pre-commit-config.yaml: v0.1.6 → v0.8.4 (matches requirements-dev.txt)

Constitutional Compliance: 52% → 70% (estimated)
Auditor: GitHub Copilot (Claude Opus 4.5)

+++ tensornet/defense/ballistics.py
@@ -111,7 +111,7 @@ class BallisticSolver:
         
         # Atmospheric conditions
         self.air_density = 1.225  # kg/m³ (sea level standard)
-        self.gravity = torch.tensor([0.0, -9.81, 0.0], device=self.device)
+        self.gravity = torch.tensor([0.0, -9.81, 0.0], device=self.device, dtype=torch.float64)
         
         # Bullet reference area (approximate for .338 caliber)
         self.bullet_diameter = 0.00861  # meters (8.61mm)
